+++ changelog.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Remove pipeline's from check (`Test-AdfCode`) for dashes in name #309
 * Bug #311: Fixed deletion of credential type of objects
 * Bug #331: Fixed `Test-AdfCode` for [credential] objects
+* Test-AdfLinkedService can use internally Invoke-AzRestMethod #144
+* Test-AdfLinkedService return an object #165
+* Test-AdfLinkedService accepts parameters of testing LS #222
 
 ## [1.4.0] - 2023-04-26
 ### Added